+++ src/CommonRuntime/TextRuntime.fs
@@ -88,7 +88,7 @@ type TextRuntime =
 
   static member ConvertDateTimeBack(cultureStr, value:DateTime option) = 
     match value with
-    | Some value -> value.ToString(TextRuntime.GetCulture cultureStr)
+    | Some value -> value.ToString("O", TextRuntime.GetCulture cultureStr)
     | None -> ""
 
   static member ConvertGuidBack(value:Guid option) =